Adicionado tela plp fechamento

+++ src/Nop.Plugin.Shipping.Correios/Models/PLPFechadaResultModel.cs
@@ -1,10 +1,6 @@
 ﻿using Nop.Web.Framework.Mvc;
 using System;
-using System.Collections.Generic;
 using System.ComponentModel;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Nop.Plugin.Shipping.Correios.Models
 {
@@ -14,7 +10,7 @@ namespace Nop.Plugin.Shipping.Correios.Models
         public int Id { get; set; }
 
         [DisplayName("PLP ID Correios")]
-        public int CorreiosId { get; set; }
+        public long CorreiosId { get; set; }
 
         [DisplayName("Data Fechamento")]
         public DateTime DataFechamento { get; set; }
